chore(condo): DOMA-10785 add component with the text of the agreement + corrected ru translation
