+++ apps/condo/domains/user/components/auth/SignInForm.tsx
@@ -19,6 +19,7 @@ import { ResponsiveCol } from '@condo/domains/user/components/containers/Respons
 import { WRONG_PASSWORD_ERROR, WRONG_PHONE_ERROR } from '@condo/domains/user/constants/errors'
 import { SIGNIN_BY_PHONE_AND_PASSWORD_MUTATION } from '@condo/domains/user/gql'
 
+import { AgreementText } from './AgreementText'
 
 const ROW_STYLES: React.CSSProperties = {
     justifyContent: 'center',
@@ -27,7 +28,7 @@ const FORM_TYPOGRAPHY_STYLES: React.CSSProperties = {
     textAlign:'center',
 }
 const FORM_PARAGRAPH_STYLES: React.CSSProperties = {
-    margin: '24px 0 40px',
+    margin: '24px 0 0',
 }
 const FORM_PHONE_STYLES: React.CSSProperties = { borderRadius: 8, borderColor: colors.inputBorderGrey }
 const FORM_BUTTONS_GUTTER: RowProps['gutter'] = [0, 20]
@@ -147,9 +148,11 @@ export const SignInForm = (): React.ReactElement => {
                                 />
                             </Typography.Paragraph>
                         </Col>
+
+                        <AgreementText />
                     </Row>
                     <Row gutter={FORM_BUTTONS_GUTTER}>
-                        <Col span={24}>
+                        <Col span={24} style={FORM_PARAGRAPH_STYLES}>
                             <Form.Item>
                                 <Button
                                     key='submit'
